split off the cff-version specific schemaorg_author behavior into its own files
